src/requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
audioplayer==0.6
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
comtypes==1.1.11
constantly==15.1.0
contourpy==1.0.5
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
distlib==0.3.6
Django==4.1.1
docutils==0.19
fastapi==0.78.0
ffpyplayer==4.3.5
filelock==3.7.1
fonttools==4.37.3
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.1
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
KivyOnTop==1.4
kiwisolver==1.4.4
lingua-franca==0.4.3
lxml==4.9.1
matplotlib==3.6.0
MouseInfo==0.1.3
numpy==1.23.0
packaging==21.3
parsel==1.6.0
Pillow==9.2.0
platformdirs==2.5.2
plyer==2.0.0
Protego==0.2.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.9.1
PyDispatcher==2.0.6
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
pystray==0.19.4
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.4
pywin32==304
pywinauto==0.6.8
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sounddevice==0.4.4
SoundFile==0.10.3.post1
sqlparse==0.4.3
srt==3.5.2
starlette==0.19.1
termcolor==1.1.0
tldextract==3.3.1
tqdm==4.64.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
typing_extensions==4.2.0
tzdata==2022.2
urllib3==1.26.9
uvicorn==0.18.2
virtualenv==20.16.5
vosk==0.3.42
w3lib==1.22.0
websockets==10.3
Wikipedia-API==0.5.4
zope.interface==5.4.0
